Remove HexArray type, use the TryFrom implementation for Hex
This removes the HexArray type, instead using the DeserializeAs
implementation for the Hex type. The DeserializeAs implementation is
strictly more general as there is a blanked implementation of TryFrom
for all From conversions.

The TryFrom<Vec<T>> for arrays is only available in Rust 1.48, thus the
test code needs to be gated on that version.