Automatically add `#[serde(borrow)]` when `BorrowCow` is used
This is a convenience feature and works since `BorrowCow` is only
implemented for types which can borrow from the input.

The behaviour can be prohibited, by providing a borrow annotation
manually. Duplicate annotations would lead to an error, so the macro
will not do that.
The borrow annotation could be wrong, if the field type has multiple
lifetime parameters, but not all should borrow.