Remove default features from chrono to get rid of the oldtime feature
time v0.1 has a security advisory RUSTSEC-2020-0071
time is an optional dependency of chrono with the default features. The
dependency is not actually needed and can be avoided by not using the
default-features. Any crates which needs that time and chrono share the
Duration type can enable the feature themself.

Closes #378