Cleanup old implementations from the `rust` module
Many implementations inside `serde_with::rust` can be better expressed
with the `#[serde_as]` macro. This commit removes these implementations.
The `StringWithSeparator` type is moved to the crate root, while the
separator implementations are moved into the `formats` module.

Not all implementations have replacements. Some cannot be used in a
nested position.

All deleted implementations are listed with replacement option in the
changelog.