Add a lockfile to keep the CI more predictable
The `time` crate bumped its MSRV beyond 1.60 what we currently use. In
the future they even plan on using stable - 2.
https://github.com/time-rs/time/discussions/535

This might require a bit too many changes. Instead with the lockfile we
use a predictable set of dependencies. MSRV bumps of dependencies will
only be noticed when we update the lockfile.