Add `KeyValueMap` type to serialize a `Vec` as a map with a field as the key
This allows converting a map into a vec. The important part is that the
map key becomes a field of the vec element.
This can be useful when the map is keyed by an ID, but the ID should end
up as part of the struct.

A JSON like this
```json
{
	"id0": {
		"field_a": ...,
		...
	},
	"id1": {...},
	"id2": {...},
}
```

can be mapped to a Rust struct like this
```rust
struct Data {
	id: IdType,
	field_a: ...,
	field_b: ...,
}
```

Closes #341