Simplify end-of-map detection for KeyValueMap
The code already uses buffering for the deserialization of the
special-key value. This allows to pull up the end-of-map check, by
buffering earlier.
This combines all the lines where the buffering happened before and puts
the end-of-map check at the correct place. Finally, we can get rid of
END_OF_MAP_IDENTIFIER.