Change documentation of feature flags
Document the feature flags in the Cargo.toml file instead of a separate
file. Keep the old file, but generate the content from the Cargo.toml
using the `document-features` crate.

The benefit of this approach is that the documentation is stored next to
the features. lib.rs has a new feature that reads the documentation from
there and shows it as part of the feature list.