Replace trybuild tests with some local ones
As it turns out, rustc's "trait unimplemented" error messages have not
stayed perfectly stable across the last 10 versions. This is likely an
issue that would keep happening with trybuild which would make it
somewhat unsuitable.

With some thinking I was able to get the same test coverage without
requiring a compile error.