Add a pattern key for the duration string
The string representation of the date actually has only a restricted set
of values that can be used. By using the pattern property we can have
the json schema itself validate the string.

The regex used here for the pattern is a more lenient version of that
used to match numbers in JSON.