Add support for schemars v0.9
This is mostly a copy-paste from the support for schemars v0.8. There
are a couple of changes which make the v0.9 support a bit more messy,
but otherwise it's not bad. This commit also duplicates the existing
tests, which all pass. Once schemars v1.0 gets released it should
hopefully require fewer changes to add support again.

The big changes in schemars that cascade out to support here are
- Schema is now just a very thin wrapper around serde_json::Value. This
  is rather convenient for defining schemas, less convenient for the
  transformations we need to make here.
- Some of the trait methods on the JsonSchema trait have had their
  definitions change.

It's possible the schemars::transform module could be used to make this
cleaner. I haven't looked too deeply into what that would involve.