Remove dependency on serde_derive
serde_derive was used in the flattened_maybe macro. It was used because
the macro needed access to a buffered deserializer for the flattened
representation. serde_with has since added the Content type from serde,
such that relying on private serde APIs is no longer necessary.

This commit removes serde_derive by inlining the macro expansion and
cleaning it up. DeserializeSeed is used to pass in the field name of the
non-flattened field into the deserialization code. This allows most of
the code to be part of serde_with and flattned_maybe! macro only needs
to invoke the deserializer.

Some extra code is necessary to handle the flattened Content buffer.